+++ logic/match/ui/__init__.py
@@ -1,0 +1,1 @@
+from .match_view import MatchJoinView, MatchResultSelect, MatchResultView